requirements.txt
llama-index==0.11.16
llama-index-embeddings-huggingface==0.3.1
llama-index-vector-stores-chroma==0.2.1
chromadb==0.5.5
sentence-transformers==3.0.1
torch>=2.2.0
transformers>=4.41.0
numpy>=1.26.0
python-dotenv>=1.0.1
openai>=1.42.0
tiktoken>=0.7.0
matplotlib
rank-bm25>=0.2.2
aiohttp>=3.9.0
asyncio>=3.4.3
diskcache>=5.6.3
scikit-learn>=1.3.0
fastapi>=0.111.0
uvicorn>=0.29.0
starlette>=0.37.2
